Added point about building native image with --no-fallback
Toolkit will be great when building CLI applications as it adds pretty much all the functionality you might need. Native image is great for building CLI applications because of the quick startup. It does however not work with CE without the --no-fallback flag. To stop users being confused by the image not working with a not obvious error message I figure it would be good to add a quick little info at the bottom here. Can probably be moved to a separate page when the docs are more fleshed out. The text i stolen from the CE documentation https://typelevel.org/cats-effect/docs/core/native-image